Move state instance and bot instance in utils.py
