+++ bot_top_ranking/utils.py
@@ -1,0 +1,10 @@
+import os
+
+import telebot
+from dotenv import load_dotenv
+
+from bot_top_ranking.config_class import State
+
+load_dotenv()
+bot = telebot.TeleBot(os.getenv("BOT_TOKEN"))
+state = State()
