+++ pytests/bot/test_decorators.py
@@ -4,32 +4,18 @@ import pytest
 from telebot import types
 
 from bot_top_ranking import decorators
-from bot_top_ranking.config_class import State
-from pytests.conftest import bot
+from bot_top_ranking.utils import state, bot
 
 User = namedtuple('User', ['user'])
-state = State()
-
-
-@pytest.mark.smoke
-@pytest.mark.decorator
-def test_get_state(mocker):
-    mocker.patch.object(bot, "get_chat_administrators", return_valus=[])
-    expected_result = (state, bot)
-
-    @decorators.get_state(*expected_result)
-    def check_params():
-        return decorators.get_state.state, decorators.get_state.bot
-    assert check_params() == expected_result
 
 
 @pytest.mark.decorator
 def test_only_admin_decorator(mocker, mock_message, capsys):
-    mocker.patch.object(bot, "get_chat_administrators", return_value=[User(types.User(bot.get_me().id, None, 'Tester'))])
+    mocker.patch.object(bot, "get_chat_administrators",
+                        return_value=[User(types.User(bot.get_me().id, None, 'Tester'))])
     expected_output = "only_admin_achieved"
 
     @decorators.only_admins
-    @decorators.get_state(state, bot)
     def check_is_admin(message):
         bot.send_message(0, expected_output)
 
@@ -44,7 +30,6 @@ def test_only_admin_decorator_raises(mocker, mock_message, capsys):
     mock_message.from_user.id = 0
 
     @decorators.only_admins
-    @decorators.get_state(state, bot)
     def stub(message):
         pass
 
@@ -59,7 +44,6 @@ def test_started_pool_decorator(mock_message, capsys):
     state.config["poll_started"] = True
 
     @decorators.started_pool
-    @decorators.get_state(state, bot)
     def check_is_started(message):
         bot.send_message(0, expected_output)
 
@@ -74,7 +58,6 @@ def test_started_pool_decorator_raises(mock_message, capsys):
     state.config["poll_started"] = False
 
     @decorators.started_pool
-    @decorators.get_state(state, bot)
     def stub(message):
         pass
